refactor: RandomNumber 클래스 추가 외 파일 경로 변경
- 외부 패키지인 RandomNumber를 분리
- PrizeAmount 상수 폴더로 변경
- 외 변수명 수정
